+++ plom/baseMessenger.py
@@ -1,9 +1,10 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2018-2020 Andrew Rechnitzer
-# Copyright (C) 2019-2022 Colin B. Macdonald
+# Copyright (C) 2019-2023 Colin B. Macdonald
 # Copyright (C) 2021 Peter Lee
 # Copyright (C) 2022 Michael Deakin
 # Copyright (C) 2022-2023 Edith Coates
+# Copyright (C) 2023 Tam Nguyen
 
 from io import BytesIO
 import logging
@@ -28,6 +29,7 @@ from plom.plom_exceptions import (
     PlomNoMoreException,
     PlomNoPaper,
     PlomNoSolutionException,
+    PlomRangeException,
     PlomServerNotReady,
     PlomSSLError,
     PlomTakenException,
@@ -764,11 +766,11 @@ class BaseMessenger:
         finally:
             self.SRmutex.release()
 
-    def MgetRubricsByQuestion(self, question_number):
+    def MgetRubricsByQuestion(self, question):
         """Retrieve list of all rubrics from server for given question.
 
         Args:
-            question_number (int)
+            question (int)
 
         Raises:
             PlomAuthenticationException: Authentication error.
@@ -781,7 +783,7 @@ class BaseMessenger:
         self.SRmutex.acquire()
         try:
             response = self.get(
-                f"/MK/rubric/{question_number}",
+                f"/MK/rubric/{question}",
                 json={
                     "user": self.user,
                     "token": self.token,
@@ -1040,7 +1042,9 @@ class BaseMessenger:
             response.raise_for_status()
             return response.json()
         except requests.HTTPError as e:
-            if response.status_code == 401:
+            if response.status_code == 400:
+                raise PlomRangeException(response.reason) from None
+            elif response.status_code == 401:
                 raise PlomAuthenticationException() from None
             elif response.status_code == 404:
                 raise PlomNoPaper(
@@ -1054,6 +1058,8 @@ class BaseMessenger:
                 raise PlomTaskDeletedError(
                     "Task {} has been deleted by manager.".format(num)
                 ) from None
+            elif response.status_code == 416:
+                raise PlomRangeException(response.reason) from None
             raise PlomSeriousException(f"Some other sort of error {e}") from None
         finally:
             self.SRmutex.release()
@@ -1086,7 +1092,9 @@ class BaseMessenger:
             response.raise_for_status()
             return BytesIO(response.content).getvalue()
         except requests.HTTPError as e:
-            if response.status_code == 401:
+            if response.status_code == 400:
+                raise PlomRangeException(response.reason) from None
+            elif response.status_code == 401:
                 raise PlomAuthenticationException() from None
             elif response.status_code == 404:
                 raise PlomNoPaper(
@@ -1100,6 +1108,8 @@ class BaseMessenger:
                 raise PlomTaskDeletedError(
                     "Task {} has been deleted by manager.".format(num)
                 ) from None
+            elif response.status_code == 416:
+                raise PlomRangeException(response.reason) from None
             raise PlomSeriousException(f"Some other sort of error {e}") from None
         finally:
             self.SRmutex.release()